+++ functions/Get-PasswordStatePasswords.ps1
@@ -17,18 +17,30 @@ function Get-PasswordStatePasswords {
     )]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingUserNameAndPassWordParams', '', Justification = 'PasswordID isnt a password')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseSingularNouns', '', Justification = 'Only returns multiple')]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidGlobalVars', '', Justification = 'Needed for backward compatability')]
     param (
-        [Parameter(ParameterSetName='GetAllPasswordsFromList', Mandatory = $false,ValueFromPipeline=$True,ValueFromPipelinebyPropertyName=$True)][int32[]]$PasswordlistID
+        [Parameter(ParameterSetName='GetAllPasswordsFromList', Mandatory = $false,ValueFromPipeline=$True,ValueFromPipelinebyPropertyName=$True, Position = 0)][int32[]]$PasswordlistID
     )
 
     begin {
+        . "$PSScriptRoot\PasswordstateClass.ps1"
+        $output = @()
     }
 
     process {
-        $output = Get-PasswordStateResource -uri $("/api/passwords/" + $PasswordlistID + "?QueryAll")
+        [PasswordResult]$results = Get-PasswordStateResource -uri $("/api/passwords/" + $PasswordlistID + "?QueryAll")
+        Foreach ($result in $results){
+            $result.Password = [EncryptedPassword]$result.Password
+            $output += $result
+        }
     }
 
     end {
-        return $output
+        switch ($global:PasswordStateShowPasswordsPlainText) {
+            True {
+                $output.DecryptPassword()
+            }
+        }
+        Return $output
     }
 }